% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Create a summary table with the first 6 rows of data}
\usage{
create_summary_table(data, path)
}
\arguments{
\item{data}{A data frame containing the data to summarize}

\item{path}{A string specifying the file path where the CSV will be saved}
}
\value{
The head of the data frame (first 6 rows) invisibly
}
\description{
This function takes a data frame and creates a summary table by selecting
the first 6 rows. It then writes this summary to a CSV file
at the specified path.
}
\examples{
\dontrun{
# Create a summary table of the knowledge_train_data
create_summary_table(knowledge_train_data, "results/table1.csv")
}
}
