% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_column.R
\name{rename_column}
\alias{rename_column}
\title{Rename a column in a data frame}
\usage{
rename_column(df, current_column_name, new_column_name)
}
\arguments{
\item{df}{A data frame containing the column to be renamed}

\item{current_column_name}{A string specifying the current name of the column}

\item{new_column_name}{A string specifying the new name for the column}
}
\value{
A data frame with the specified column renamed
}
\description{
This function renames a column in a data frame using dplyr's rename function
with non-standard evaluation through rlang's sym function.
}
\examples{
\dontrun{
# Create a sample data frame
df <- data.frame(
  old_name = c(1, 2, 3),
  other_col = c("A", "B", "C")
)

# Rename the 'old_name' column to 'new_name'
df_renamed <- rename_column(df, "old_name", "new_name")
}
}
